dre4m-backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
autopep8==2.3.2
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
flake8==7.1.2
git-filter-repo==2.47.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pytest==8.3.5
pytest-order==1.3.0
pytest-ordering==0.6
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomli==2.2.1
typer==0.15.1
typing-inspection==0.4.0
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
